refactor(view): extract motion_geometry.cpp from motion.cpp
The Layout/Presentation geometry walker + scroll-geometry helpers
relocated from motion.cpp into core/view/src/motion_geometry.cpp
(~299 lines): 2D affine matrix math, local→global View transforms,
axis-aligned bounding boxes, and the geometry entry points the motion
Coordinator consumes.

The cross-section interface is small and one-directional — the
geometry block has zero back-dependency on the Coordinator. Six entry
points (resolve_geometry, extract_property, property_name,
extract_scroll_property, scroll_property_name, fmt_double) are declared
in a new private core/view/src/motion_geometry_internal.hpp; the
file-local matrix/AABB helpers stay private to the new TU.

motion.cpp: 2,162 → 1,881 (-281). Geometry-walker work no longer
recompiles the 2.1k-line Coordinator.

Verified: pulp-view builds clean (with Skia); pulp-test-motion passes
(199 assertions / 57 test cases).
